Update Bootstrap, Model,SalesOrder,purchase,Bill

+++ Controllers/PurchaseOrdersController.cs
@@ -7,6 +7,9 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using IMS.Data;
 using IMS.Models;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
+using Microsoft.CodeAnalysis;
 
 namespace IMS.Controllers
 {
@@ -18,11 +21,11 @@ namespace IMS.Controllers
         {
             _context = context;
         }
-
+      
         // GET: PurchaseOrders
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.PurchaseOrder.Include(p => p.Branch).Include(p => p.Product).Include(p => p.Supplier);
+            var applicationDbContext = _context.PurchaseOrder.Include(p => p.Bill).Include(p => p.Branch).Include(p => p.Product).Include(p => p.Supplier);
             return View(await applicationDbContext.ToListAsync());
         }
 
@@ -50,27 +53,35 @@ namespace IMS.Controllers
         // GET: PurchaseOrders/Create
         public IActionResult Create()
         {
-            ViewData["BranchId"] = new SelectList(_context.Branch, "BranchId", "BranchName");
-            ViewData["ProductId"] = new SelectList(_context.Product, "ProductId", "ProductName");
+            ViewData["BranchId"] = new SelectList(_context.Branch, "BranchId", "BranchId");
+            ViewData["ProductId"] = new SelectList(_context.Product, "ProductId", "ProductCode");
             ViewData["SupplierId"] = new SelectList(_context.Supplier, "SupplierId", "SupplierName");
             return View();
         }
-
+      
         // POST: PurchaseOrders/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("PurchaseOrderId,SupplierId,BranchId,ProductId,OrderDate,DeliveryDate,Price,Quantity,Total,Remarks")] PurchaseOrder purchaseOrder)
+        public async Task<IActionResult> Create([Bind("BranchId,SupplierId,ProductId,Quantity,Price,Amount,OrderDate,DeliveryDate,Remarks,BillId")] PurchaseOrder purchaseOrder)
         {
+            var user = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            purchaseOrder.AdminId = user;
             if (ModelState.IsValid)
             {
+
                 _context.Add(purchaseOrder);
+                var UpdateProduct = await _context.Product.FindAsync(purchaseOrder.ProductId);
+                if (UpdateProduct == null)
+                    return NotFound();
+                UpdateProduct.CurrentQuantity += purchaseOrder.Quantity;
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["BranchId"] = new SelectList(_context.Branch, "BranchId", "BranchName", purchaseOrder.BranchId);
-            ViewData["ProductId"] = new SelectList(_context.Product, "ProductId", "ProductName", purchaseOrder.ProductId);
+            ViewData["BranchId"] = new SelectList(_context.Branch, "BranchId", "BranchId", purchaseOrder.BranchId);
+            ViewData["ProductId"] = new SelectList(_context.Product, "ProductId", "ProductCode", purchaseOrder.ProductId);
             ViewData["SupplierId"] = new SelectList(_context.Supplier, "SupplierId", "SupplierName", purchaseOrder.SupplierId);
             return View(purchaseOrder);
         }
@@ -88,8 +99,9 @@ namespace IMS.Controllers
             {
                 return NotFound();
             }
-            ViewData["BranchId"] = new SelectList(_context.Branch, "BranchId", "BranchName", purchaseOrder.BranchId);
-            ViewData["ProductId"] = new SelectList(_context.Product, "ProductId", "ProductName", purchaseOrder.ProductId);
+            ViewData["BillId"] = new SelectList(_context.Set<Bill>(), "BillId", "BillId", purchaseOrder.BillId);
+            ViewData["BranchId"] = new SelectList(_context.Branch, "BranchId", "BranchId", purchaseOrder.BranchId);
+            ViewData["ProductId"] = new SelectList(_context.Product, "ProductId", "ProductCode", purchaseOrder.ProductId);
             ViewData["SupplierId"] = new SelectList(_context.Supplier, "SupplierId", "SupplierName", purchaseOrder.SupplierId);
             return View(purchaseOrder);
         }
@@ -99,23 +111,63 @@ namespace IMS.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("PurchaseOrderId,SupplierId,BranchId,ProductId,OrderDate,DeliveryDate,Price,Quantity,Total,Remarks")] PurchaseOrder purchaseOrder)
+        public async Task<IActionResult> Edit(int PurchaseOrderId,int BranchId,int SupplierId,int ProductId,int Quantity,double Price,double Amount, DateTimeOffset OrderDate, DateTimeOffset DeliveryDate,string? Remarks,string? BillId,string AdminId)
         {
-            if (id != purchaseOrder.PurchaseOrderId)
-            {
+            if(PurchaseOrderId == null)
                 return NotFound();
-            }
 
+            //AdminId= User.FindFirstValue(ClaimTypes.NameIdentifier);
             if (ModelState.IsValid)
             {
+                PurchaseOrder purchase = _context.PurchaseOrder.FirstOrDefault(item => item.PurchaseOrderId == PurchaseOrderId);
+                if (purchase == null)
+                    return NotFound();
+                if (purchase.ProductId == ProductId)
+                {
+                    var UpdateProduct = await _context.Product.FindAsync(ProductId);
+                    if (UpdateProduct == null)
+                        return NotFound();
+                    if (purchase.Quantity > Quantity)
+                    {
+                        var decreaseQuantity = purchase.Quantity - Quantity;
+                        UpdateProduct.CurrentQuantity -= decreaseQuantity;
+                    }
+                    if (purchase.Quantity < Quantity)
+                    {
+                        var increaseQuantity = Quantity - purchase.Quantity;
+                        UpdateProduct.CurrentQuantity += increaseQuantity;
+                    }
+
+                }
+                else
+                {
+                    var UpdateProduct = await _context.Product.FindAsync(purchase.ProductId);
+                    UpdateProduct.CurrentQuantity-=purchase.Quantity;
+
+                    var NewUpdateProduct = await _context.Product.FindAsync(ProductId);
+                    NewUpdateProduct.CurrentQuantity += Quantity;
+
+
+                }
+
+                purchase.BranchId = BranchId;
+                purchase.SupplierId = SupplierId;
+                purchase.ProductId = ProductId;
+                purchase.Quantity = Quantity;
+                purchase.Price = Price;
+                purchase.Amount = Amount;
+                purchase.OrderDate = OrderDate;
+                purchase.DeliveryDate = DeliveryDate;
+                purchase.Remarks = Remarks;
+                purchase.AdminId = AdminId;
+
                 try
                 {
-                    _context.Update(purchaseOrder);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!PurchaseOrderExists(purchaseOrder.PurchaseOrderId))
+                    if (!PurchaseOrderExists(purchase.PurchaseOrderId))
                     {
                         return NotFound();
                     }
@@ -126,10 +178,23 @@ namespace IMS.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["BranchId"] = new SelectList(_context.Branch, "BranchId", "BranchName", purchaseOrder.BranchId);
-            ViewData["ProductId"] = new SelectList(_context.Product, "ProductId", "ProductName", purchaseOrder.ProductId);
-            ViewData["SupplierId"] = new SelectList(_context.Supplier, "SupplierId", "SupplierName", purchaseOrder.SupplierId);
-            return View(purchaseOrder);
+            PurchaseOrder po = new PurchaseOrder();
+            po.PurchaseOrderId = PurchaseOrderId;
+            po.BranchId = BranchId;
+            po.SupplierId = SupplierId;
+            po.ProductId = ProductId;
+            po.Quantity = Quantity;
+            po.Price = Price;
+            po.Amount = Amount;
+            po.OrderDate = OrderDate;
+            po.DeliveryDate = DeliveryDate;
+            po.Remarks = Remarks;
+            po.AdminId = AdminId;
+            ViewData["BillId"] = new SelectList(_context.Set<Bill>(), "BillId", "BillId", po.BillId);
+            ViewData["BranchId"] = new SelectList(_context.Branch, "BranchId", "BranchId", po.BranchId);
+            ViewData["ProductId"] = new SelectList(_context.Product, "ProductId", "ProductCode", po.ProductId);
+            ViewData["SupplierId"] = new SelectList(_context.Supplier, "SupplierId", "SupplierName", po.SupplierId);
+            return View(po);
         }
 
         // GET: PurchaseOrders/Delete/5
@@ -141,6 +206,7 @@ namespace IMS.Controllers
             }
 
             var purchaseOrder = await _context.PurchaseOrder
+                .Include(p => p.Bill)
                 .Include(p => p.Branch)
                 .Include(p => p.Product)
                 .Include(p => p.Supplier)
@@ -165,7 +231,13 @@ namespace IMS.Controllers
             var purchaseOrder = await _context.PurchaseOrder.FindAsync(id);
             if (purchaseOrder != null)
             {
-                _context.PurchaseOrder.Remove(purchaseOrder);
+                var UpdateProduct = await _context.Product.FindAsync(purchaseOrder.ProductId);
+                if (UpdateProduct == null)
+                    return NotFound();
+                {
+                    UpdateProduct.CurrentQuantity -= purchaseOrder.Quantity;
+                }
+                _context.PurchaseOrder.Remove(purchaseOrder);      
             }
             
             await _context.SaveChangesAsync();
